changed the type of percentage in rapl.Cap(...) from int to float64. Retrofitted power-capping strategies to cap using a float64 value instead of an int. Moved common functions in loganddynamiccap.go and logAndProgressiveExtrema.go into pcp/utils.go. New power-capping strategy that builds on top of extrema, where it caps the victims at different until it can't cap further, in which case it starts uncapping them in the reverse order of capping.

+++ schedulers/bpswMaxMinPistonCapping.go
@@ -157,7 +157,7 @@ var bpMaxMinPistonCappingMutex sync.Mutex
 var bpMaxMinPistonCappingCapValues = make(map[string]float64)
 
 // Storing the previous cap value for each host so as to not repeatedly cap the nodes to the same value. (reduces overhead)
-var bpMaxMinPistonCappingPreviousRoundedCapValues = make(map[string]int)
+var bpMaxMinPistonCappingPreviousRoundedCapValues = make(map[string]float64)
 
 func (s *BPSWMaxMinPistonCapping) startCapping() {
 	go func() {
@@ -167,7 +167,7 @@ func (s *BPSWMaxMinPistonCapping) startCapping() {
 				// Need to cap each node
 				bpMaxMinPistonCappingMutex.Lock()
 				for host, capValue := range bpMaxMinPistonCappingCapValues {
-					roundedCapValue := int(math.Floor(capValue + 0.5))
+					roundedCapValue := float64(int(math.Floor(capValue + 0.5)))
 					// has the cap value changed
 					if previousRoundedCap, ok := bpMaxMinPistonCappingPreviousRoundedCapValues[host]; ok {
 						if previousRoundedCap != roundedCapValue {
